% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcluster.R
\name{addcluster}
\alias{addcluster}
\title{Function that adds the clustering result to a \code{seqimp} object
obtained with the \code{seqimpute} function}
\usage{
addcluster(impdata, clustering)
}
\arguments{
\item{impdata}{An object of class \code{seqimp} as created by the 
\code{seqimpute} function}

\item{clustering}{clustering made on the multiple imputed dataset. Can 
either be a dataframe or a matrix, where each row correspond to an
observation and each column to a multiple imputed dataset}
}
\value{
Returns a \code{seqimp} object containing the cluster to which each 
sequence in each imputed dataset belongs. Specifically, a column named 
cluster is added to the imputed datasets.
}
\description{
Function that adds the clustering result to a \code{seqimp} object
obtained with the \code{seqimpute} function
}
