% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI3_SpecificModelImputation.R
\name{CDCompute}
\alias{CDCompute}
\title{Building of the data matrix CD for the computation of the model}
\usage{
CDCompute(
  CO,
  OD,
  COt,
  MaxGap,
  order,
  np,
  nc,
  nr,
  nf,
  COtsample,
  pastDistrib,
  futureDistrib,
  ncot,
  nco,
  k,
  timing
)
}
\description{
Building of the data matrix CD for the computation of the model
}
