% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI4_InitialNAsImpute.R
\name{CDMatCreate}
\alias{CDMatCreate}
\title{Building of the data matrix CD for the computation of the model}
\usage{
CDMatCreate(CO, COtsample, OD, COt, nfi, nr, nc, ncot, futureDistrib, k)
}
\description{
Building of the data matrix CD for the computation of the model
}
