% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{CO}
\alias{CO}
\title{Dataset containing 3 fixed covariates about the game addiction of young subjects}
\format{
A data frame of fixed covariates with 500 sequences and 3 columns.
}
\usage{
data(CO)
}
\description{
These covariates are respectively '\code{Gender}' (\code{male}/\code{female}), '\code{Age}' (at \code{T1}, in years) and '\code{Track}' (\code{school}/\code{apprenticeship}).
}
\details{
\itemize{
  \item 500 sequences
  \item 3 columns
}
}
\keyword{datasets}
