% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICommon.R
\name{ComputeModel}
\alias{ComputeModel}
\title{Compute model with the chosen regression model}
\usage{
ComputeModel(
  CD,
  regr,
  tot_VI,
  np,
  nf,
  k,
  num.trees,
  min.node.size,
  max.depth,
  timing
)
}
\description{
Compute model with the chosen regression model
}
