% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{COt}
\alias{COt}
\title{Dataset containing 1 time-dependant covariate about the game addiction of young subjects}
\format{A data frame of time-dependant covariates with 500 sequences and 4 columns.}
\usage{
data(COt)
}
\description{
This time-dependant covariate is the '\code{Gambling}' (\code{no}/\code{gambler}/\code{problematic gambler}) and contains thus the same number of columns as the original data frame \code{OD}: 4 columns.
}
\details{
\itemize{
  \item 500 sequences
  \item 4 columns
}
}
\keyword{datasets}
