% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICommon.R
\name{COtselectionSpe}
\alias{COtselectionSpe}
\title{Concatenating CDi with COtselected_i (the matrix containing the current 
time-dependent covariates) Checking if COt is NOT completely empty}
\usage{
COtselectionSpe(CDi, COt, ncot, nc, i, j, k)
}
\description{
Concatenating CDi with COtselected_i (the matrix containing the current 
time-dependent covariates) Checking if COt is NOT completely empty
}
