% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI3_SpecificModelImputation.R
\name{CreatedModelImputation}
\alias{CreatedModelImputation}
\title{Imputation using the just created model (Dealing with the actual VALUES to impute)}
\usage{
CreatedModelImputation(
  order,
  CO,
  CD,
  COt,
  OD,
  ODi,
  pastDistrib,
  futureDistrib,
  available,
  REFORD_L,
  ncot,
  nc,
  np,
  nf,
  k,
  totV,
  regr,
  reglog,
  noise,
  shift,
  MaxGap
)
}
\description{
Imputation using the just created model (Dealing with the actual VALUES to impute)
}
