% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI0_InitCheck.R
\name{defineNbVariables}
\alias{defineNbVariables}
\title{Definition of the number of rows and columns in OD
(And eventually update of nr)}
\usage{
defineNbVariables(
  OD,
  COt,
  ncot,
  np,
  nco,
  nf,
  nfi,
  npt,
  k,
  pastDistrib,
  futureDistrib
)
}
\description{
Definition of the number of rows and columns in OD
(And eventually update of nr)
}
