% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI0_InitCheck.R
\name{deleteNaRows}
\alias{deleteNaRows}
\title{Deleting entire rows of OD filled only with NAs
(and deleting corresponding lines in CO and COt)}
\usage{
deleteNaRows(OD, CO, COt)
}
\description{
Deleting entire rows of OD filled only with NAs
(and deleting corresponding lines in CO and COt)
}
