% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI0_InitCheck.R
\name{factorAndMatrixChecks}
\alias{factorAndMatrixChecks}
\title{In case of a factor dataset OD:
RECODING of OD with numbers "1", "2", etc. instead of its "words"}
\usage{
factorAndMatrixChecks(OD, nc, k)
}
\description{
In case of a factor dataset OD:
RECODING of OD with numbers "1", "2", etc. instead of its "words"
}
