% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI2_ComputeImputeOrder.R
\name{FutObsCompute}
\alias{FutObsCompute}
\title{Model 3: use of future observations only}
\usage{
FutObsCompute(ORDER, ORDER3, np, nf, nr, nc, MaxGap)
}
\description{
Model 3: use of future observations only
}
