% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI1_ORDERCreation.R
\name{GapSizeCreation}
\alias{GapSizeCreation}
\title{Creation of Gapsize}
\usage{
GapSizeCreation(ORDER, nr, OrderWidth, OrderList)
}
\description{
Creation of vector InitGapSize (i.e. a vector containing the size of the
initial gaps of each line)
}
