% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI2_ComputeImputeOrder.R
\name{ImputeOrderComputation}
\alias{ImputeOrderComputation}
\title{Compute the order of imputation
   
In case of a factor dataset OD:
RECODING of OD with numbers "1", "2", etc. instead of its "words"}
\usage{
ImputeOrderComputation(ORDER, ORDER3, MaxGap, np, nf, nr, nc)
}
\description{
Compute the order of imputation
   
In case of a factor dataset OD:
RECODING of OD with numbers "1", "2", etc. instead of its "words"
}
