% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI4_InitialNAsImpute.R
\name{ImputingInitialNAs}
\alias{ImputingInitialNAs}
\title{Impute initial NAs}
\usage{
ImputingInitialNAs(
  CO,
  COt,
  OD,
  ODi,
  totVi,
  COtsample,
  futureDistrib,
  InitGapSize,
  MaxInitGapSize,
  nr,
  nc,
  ud,
  ncot,
  nfi,
  regr,
  k,
  available,
  noise,
  num.trees,
  min.node.size,
  max.depth,
  timing
)
}
\description{
Impute initial NAs
}
