% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI5_TerminalNasImpute.R
\name{ImputingTerminalNAs}
\alias{ImputingTerminalNAs}
\title{Impute terminal NAs}
\usage{
ImputingTerminalNAs(
  ODi,
  CO,
  OD,
  COt,
  COtsample,
  MaxTermGapSize,
  TermGapSize,
  pastDistrib,
  regr,
  npt,
  ncot,
  totVt,
  nr,
  nc,
  ud,
  available,
  k,
  noise,
  num.trees,
  min.node.size,
  max.depth
)
}
\description{
Impute terminal NAs
}
