% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI4_InitialNAsImpute.R
\name{Init_NA_CreatedModelImputation}
\alias{Init_NA_CreatedModelImputation}
\title{Imputation using the just created model (Dealing with the actual VALUES to impute)}
\usage{
Init_NA_CreatedModelImputation(
  OD,
  ODi,
  CO,
  CD,
  COt,
  MaxInitGapSize,
  REFORDI_L,
  futureDistrib,
  totVi,
  nc,
  k,
  nfi,
  ncot,
  regr,
  reglog,
  noise,
  available
)
}
\description{
Imputation using the just created model (Dealing with the actual VALUES to impute)
}
