% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI0_InitCheck.R
\name{InitCorectControl}
\alias{InitCorectControl}
\title{Initial tests and manipulations on parameters}
\usage{
InitCorectControl(
  regr,
  ODClass,
  OD,
  nr,
  nc,
  k,
  np,
  nf,
  nco,
  ncot,
  nfi,
  npt,
  pastDistrib,
  futureDistrib,
  totV,
  totVi,
  totVt,
  noise
)
}
\description{
Initial tests and manipulations on parameters
}
