% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI6_SLGNAsImpute.R
\name{LSLGNAsImpute}
\alias{LSLGNAsImpute}
\title{Left-hand side SLG imputation}
\usage{
LSLGNAsImpute(
  OD,
  ODi,
  CO,
  COtsample,
  ORDERSLG,
  pastDistrib,
  futureDistrib,
  regr,
  np,
  nr,
  nf,
  nc,
  ud,
  ncot,
  nco,
  k,
  noise,
  available,
  num.trees,
  min.node.size,
  max.depth,
  timing
)
}
\description{
Left-hand side SLG imputation
}
