% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missings_generation.R
\name{make_missing_MAR}
\alias{make_missing_MAR}
\title{Generation of missing data under the form of gaps, which
is the typical form of missing data with longitudinal data.
A missing completely at random (MAR) mechanism is used.}
\usage{
make_missing_MAR(
  data,
  pstart_high = 0.2,
  pstart_low = 0.03,
  propdata = 0.6,
  states_high
)
}
\arguments{
\item{data}{a data frame containing sequences of a multinomial variable without missing
data, on which missing data are simulated}

\item{pstart_high}{probability to start a missing data for the specified states}

\item{pstart_low}{probability to start a missing data for the other states}

\item{propdata}{proportion of the observations on which missing data will be simulated
subsequent gap of missing data}

\item{states_high}{list of states that will have a larger probability to trigger a}
}
\description{
Generation of missing data under the form of gaps, which
is the typical form of missing data with longitudinal data.
A missing completely at random (MAR) mechanism is used.
}
