% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missings_generation.R
\name{make_missing_MCAR}
\alias{make_missing_MCAR}
\title{Generation of missing data under the form of gaps, which
is the typical form of missing data with longitudinal data.
A missing completely at random (MCAR) mechanism is used.}
\usage{
make_missing_MCAR(data, pstart = 0.1, propdata = 0.6)
}
\arguments{
\item{data}{a data frame containing sequences of a multinomial variable without missing
data, on which missing data are simulated}

\item{pstart}{probability to start a missing data}

\item{propdata}{proportion of the observations on which missing data will be simulated}
}
\description{
Generation of missing data under the form of gaps, which
is the typical form of missing data with longitudinal data.
A missing completely at random (MCAR) mechanism is used.
}
