% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI3_SpecificModelImputation.R
\name{ModelImputation}
\alias{ModelImputation}
\title{Impute data using a specific model}
\usage{
ModelImputation(
  OD,
  CO,
  COt,
  ODi,
  MaxGap,
  totV,
  totVi,
  regr,
  nc,
  np,
  nf,
  nr,
  ncot,
  COtsample,
  pastDistrib,
  futureDistrib,
  k,
  available,
  REFORD_L,
  noise,
  num.trees,
  min.node.size,
  max.depth
)
}
\description{
Impute data using a specific model
}
