% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{OD}
\alias{OD}
\title{Dataset containing variables about the game addiction of young subjects}
\format{
A data frame of factor variables with 500 sequences and 4 columns.
}
\usage{
data(OD)
}
\description{
An original dataset example \code{OD} to test \code{seqimpute.R} and its aside functions.
}
\details{
\itemize{
  \item 500 sequences (i.e. 500 rows)
  \item 4 time measurements (i.e. 4 columns)
  \item The variables can be either 'no', 'yes' or NA
}
}
\keyword{datasets}
