% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI3_SpecificModelImputation.R
\name{ODiImputeFUTURE}
\alias{ODiImputeFUTURE}
\title{Imputation where only FUTURE VIs exist}
\usage{
ODiImputeFUTURE(
  CO,
  ODi,
  CD,
  COt,
  REFORD,
  nr_REFORD,
  pastDistrib,
  futureDistrib,
  k,
  np,
  nf,
  nc,
  ncot,
  totV,
  reglog,
  LOOKUP,
  regr,
  noise
)
}
\description{
Imputation where only FUTURE VIs exist
}
