% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICommon.R
\name{ODiImputePAST}
\alias{ODiImputePAST}
\title{Imputation where only PAST VIs  exist}
\usage{
ODiImputePAST(
  CO,
  ODi,
  CD,
  COt,
  REFORD,
  nr_REFORD,
  pastDistrib,
  futureDistrib,
  k,
  np,
  nf,
  nc,
  ncot,
  totV,
  reglog,
  LOOKUP,
  regr,
  noise,
  timing
)
}
\description{
Imputation where only PAST VIs  exist
}
