% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICommon.R
\name{ODiImputePF}
\alias{ODiImputePF}
\title{Imputation where past and future VIs exist}
\usage{
ODiImputePF(
  CO,
  ODi,
  CD,
  COt,
  REFORD,
  nr_REFORD,
  pastDistrib,
  futureDistrib,
  k,
  np,
  nf,
  nc,
  ncot,
  totV,
  reglog,
  LOOKUP,
  regr,
  noise,
  shift,
  MaxGap,
  order
)
}
\description{
Imputation where past and future VIs exist
}
