% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{onlyimputed}
\alias{onlyimputed}
\title{Extract only the completed datasets from the results obtained with seqimpute function.
Therefore, the original dataset with missing values is discarded,
together with the two first columns (".id" and ".imp")".}
\usage{
onlyimputed(impdata)
}
\arguments{
\item{impdata}{a dataframe obtained with the seqimpute() function}
}
\value{
Returns a data frame that has the form for the application of
Halpin's clustering strategy
}
\description{
Extract only the completed datasets from the results obtained with seqimpute function.
Therefore, the original dataset with missing values is discarded,
together with the two first columns (".id" and ".imp")".
}
