% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI1_ORDERCreation.R
\name{OrderCreation}
\alias{OrderCreation}
\title{Creation of matrix ORDER}
\usage{
OrderCreation(OD, nr, nc)
}
\description{
Coding of missing data in function of the length of the gap
ORDER: matrix of the same size of OD giving the imputation order of each
MD (0 for observed data and 1 for MD) --> there will be 1 everywhere there
is MD and 0 everywhere else
ORDER2: matrix of the same size of OD numbering MD into each gap (for
example from 1 to 3 for a gap of length 3)
ORDER3: matrix of the same size of OD replacing each MD by the length of
the gap it belongs to.
}
