% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI2_ComputeImputeOrder.R
\name{ORDERSLGCreation}
\alias{ORDERSLGCreation}
\title{Creation of ORDERSLG (ORDERSLGLeft and ORDERSLGRight)}
\usage{
ORDERSLGCreation(ORDER, nr, nc, np, nf)
}
\description{
Updating ORDER with "0" on every NAs belonging to a Specially Located Gap 
(SLG) (The purpose of this modification of ORDER is that we don't take into
account SLG NAs at this moment of the program.
We will first impute internal gaps, external gaps and consider SLG at the 
very end (as far as some SLG have been detected)
}
