% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI2_ComputeImputeOrder.R
\name{ORDERSLGLRCompute}
\alias{ORDERSLGLRCompute}
\title{Computation of ORDERSLG (ORDERSLGLeft and ORDERSLGRight)}
\usage{
ORDERSLGLRCompute(
  nr,
  nc,
  ORDERSLG,
  tempMinGapLeft,
  tempMinGapRight,
  tempMaxGapLeft,
  tempMaxGapRight
)
}
\description{
Extracting extrema from tempMinGapLeft, tempMaxGapLeft,
tempMinGapRight and tempMaxGapRight
And creation of ORDERSLGLeft and ORDERSLGRight (matrices for both
groups of SLG (i.e. one on the left- and the other one on the right-
hand side of the matrix ORDERSLG)
}
