% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICommon.R
\name{PastFutureVICompute}
\alias{PastFutureVICompute}
\title{Past or future VI computation}
\usage{
PastFutureVICompute(
  CD,
  CO,
  OD,
  ncot,
  frameSize,
  iter,
  nr,
  nc,
  ud,
  np,
  COtsample,
  COt,
  pastDistrib,
  futureDistrib,
  k,
  nf,
  shift
)
}
\description{
Past or future VI computation
}
