% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICommon.R
\name{PastVICompute}
\alias{PastVICompute}
\title{Past VI computation}
\usage{
PastVICompute(
  CD,
  CO,
  OD,
  ncot,
  frameSize,
  iter,
  nr,
  nc,
  ud,
  np,
  COtsample,
  COt,
  pastDistrib,
  futureDistrib,
  k
)
}
\description{
Past VI computation
}
