% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.seqimp.R
\name{plot.seqimp}
\alias{plot.seqimp}
\title{Plot a \code{seqimp} object}
\usage{
\method{plot}{seqimp}(x, m = 5, include = TRUE, ...)
}
\arguments{
\item{x}{Object of class \code{seqimp}}

\item{m}{Number of completed datasets to show}

\item{include}{logical that indicates if the original dataset with missing
value should be plotted or not}

\item{...}{Arguments to be passed to the seqdplot function}
}
\description{
Plot a \code{seqimp} object. The state distribution plot of the first 
\code{m} completed datasets is shown, possibly alongside the original 
dataset with missing data
}
\author{
Kevin Emery
}
