% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI0_InitCheck.R
\name{preliminaryChecks}
\alias{preliminaryChecks}
\title{In case of a factor dataset OD:
RECODING of OD with numbers "1", "2", etc. instead of its "words"}
\usage{
preliminaryChecks(OD, CO, COt, np, nf, nfi, npt, k, pastDistrib, futureDistrib)
}
\description{
In case of a factor dataset OD:
RECODING of OD with numbers "1", "2", etc. instead of its "words"
}
