% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI2_ComputeImputeOrder.R
\name{PrevAndFutCompute}
\alias{PrevAndFutCompute}
\title{Model 1: use of previous and future observations}
\usage{
PrevAndFutCompute(ORDER, ORDER3, np, nf, nr, nc, MaxGap)
}
\description{
Model 1: use of previous and future observations
}
