% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI2_ComputeImputeOrder.R
\name{PrevObsCompute}
\alias{PrevObsCompute}
\title{Model 2: use of previous observations only}
\usage{
PrevObsCompute(ORDER, ORDER3, np, nf, nr, nc, MaxGap)
}
\description{
Model 2: use of previous observations only
}
