% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.seqimp.R
\name{print.seqimp}
\alias{print.seqimp}
\title{Print a \code{seqimp} object}
\usage{
\method{print}{seqimp}(x, ...)
}
\arguments{
\item{x}{Object of class \code{seqimp}}

\item{...}{additional arguments passed to other functions}
}
\description{
Print a \code{seqimp} object
}
\author{
Kevin Emery
}
