% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICommon.R
\name{RegrImpute}
\alias{RegrImpute}
\title{Impute value with the chosen regression model}
\usage{
RegrImpute(ODi, CDi, regr, reglog, noise, i, j, k)
}
\description{
Impute value with the chosen regression model
}
