% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{seqcomplete}
\alias{seqcomplete}
\title{Extract all the sequences without missing value.}
\usage{
seqcomplete(seqdata)
}
\arguments{
\item{seqdata}{a state sequence object built with the TraMineR package}
}
\value{
Returns either a data frame or a state sequence object, depending
the type of data that was provided to the function
}
\description{
Extract all the sequences without missing value.
}
