% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqImpTrans.R
\name{seqImpTrans}
\alias{seqImpTrans}
\title{Spotting impossible transitions among a dataset}
\usage{
seqImpTrans(OD, k, impTrans)
}
\arguments{
\item{OD}{\code{matrix} object containing sequences of a variable with missing data (coded as \code{NA}).}

\item{k}{\code{numeric} object corresponding to the number of categories of the variable numbered from \code{1} to \code{k}.}

\item{impTrans}{\code{character} vector gathering the impossible transitions. For example: impTrans <- c("1->3","1->4","2->1","4->1","4->3")}
}
\value{
It returns a message containing the impTransOverview \code{data.frame} object that gathers the occurences of each type of impossible transitions.

rowMat \code{matrix} object containing the row coordinates of the impossible transitions (stored in \code{seqImpTransList[1]}).

colMat \code{matrix} object containing the column coordinates of the impossible transitions (stored in \code{seqImpTransList[2]}).
}
\description{
The purpose of \code{seqImpTrans.R} is to spot the impossible transitions
in a dataset. These could have been introduced by mistake during data
collection or not intelligently imputed during a first run of
\code{seqimpute.R}.
}
\examples{
data(OD)

seqImpTransList <- seqImpTrans(OD=OD, k=2, impTrans=c("yes->no"))

}
\author{
Andre Berchtold <andre.berchtold@unil.ch>
}
\keyword{data,}
\keyword{impossible}
\keyword{linear}
\keyword{logistic}
\keyword{missing}
\keyword{multinomial}
\keyword{ordinal}
\keyword{regression,}
\keyword{transitions}
