% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqMissPatt.R
\name{seqMissPatt}
\alias{seqMissPatt}
\title{Visualizing missing patterns among a dataset}
\usage{
seqMissPatt(OD, k, numbOfMostFreqSeq = 10, pbarw = FALSE, clustNumb = 3)
}
\arguments{
\item{OD}{\code{matrix} object containing sequences of a variable with missing data (coded as \code{NA}).}

\item{k}{\code{numeric} object corresponding to the number of categories of the multinomial variable numbered from \code{1} to \code{k}.}

\item{numbOfMostFreqSeq}{\code{numeric} object defining the number of most frequent sequences \code{TraMineR} plots are going to display (default \code{10}).}

\item{pbarw}{\code{logical} object making the width of the bars of the \code{TraMineR} plots proportional to the sequence frequency in the dataset if \code{pbarw=TRUE} (default \code{FALSE}).}

\item{clustNumb}{\code{numeric} object corresponding to the number of clusters used to characterize the original dataset \code{OD} (default \code{3}) (min value: \code{2}, max value: \code{9}).}
}
\value{
Various plots indexing the missing patterns among the binary matrix
of the inputed dataset:
\itemize{
  \item Sequences ordered according to ascending size of initial/terminal gaps
  \item Most frequent sequences
  \item Clustered sequences
  \item Etc.
}
}
\description{
\code{seqMissPatt.R} offers a quick visualization of the repeating missing
patterns among the dataset especially thanks to the package \code{TraMineR}.
}
\examples{
data(OD, package="seqimpute")

seqMissPatt(OD=OD, k=2, numbOfMostFreqSeq=5, pbarw=TRUE, clustNumb=3)

}
\author{
Andre Berchtold <andre.berchtold@unil.ch>
}
\keyword{data,}
\keyword{linear}
\keyword{logistic}
\keyword{missing}
\keyword{multinomial}
\keyword{ordinal}
\keyword{patterns}
\keyword{regression,}
