% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI6_SLGNAsImpute.R
\name{SLGCDMatBuild}
\alias{SLGCDMatBuild}
\title{Compute the CD matrix for SLG}
\usage{
SLGCDMatBuild(
  CO,
  OD,
  ODt,
  order,
  MaxGapSLGLeft,
  np,
  ncot,
  nr,
  nc,
  nf,
  COtsample,
  pastDistrib,
  futureDistrib,
  k
)
}
\description{
Compute the CD matrix for SLG
}
