% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI6_SLGNAsImpute.R
\name{SLGCreatedModelImpute}
\alias{SLGCreatedModelImpute}
\title{Impute SLG using created model}
\usage{
SLGCreatedModelImpute(
  CO,
  OD,
  CD,
  ODi,
  COt,
  ncot,
  nf,
  nc,
  regr,
  k,
  totV_temp,
  reglog_6,
  noise,
  available,
  REFORDSLG_L,
  np,
  pastDistrib,
  futureDistrib,
  order,
  MaxGap,
  shift,
  timing
)
}
\description{
Impute SLG using created model
}
