% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.seqimp.R
\name{summary.seqimp}
\alias{summary.seqimp}
\title{Summary of a \code{seqimp} object}
\usage{
\method{summary}{seqimp}(object, ...)
}
\arguments{
\item{object}{Object of class \code{seqimp}}

\item{...}{additional arguments passed to other functions}
}
\description{
Summary of a \code{seqimp} object
}
\author{
Kevin Emery
}
