% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI5_TerminalNasImpute.R
\name{TerminalCDMatCreate}
\alias{TerminalCDMatCreate}
\title{Building of the data matrix CD for the computation of the model}
\usage{
TerminalCDMatCreate(CO, OD, COt, COtsample, pastDistrib, npt, nr, nc, ncot, k)
}
\description{
Building of the data matrix CD for the computation of the model
}
