% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI5_TerminalNasImpute.R
\name{TerminalCreatedModelImputation}
\alias{TerminalCreatedModelImputation}
\title{Imputation of the terminal created model}
\usage{
TerminalCreatedModelImputation(
  CO,
  OD,
  CD,
  ODi,
  COt,
  nc,
  ncot,
  totVt,
  REFORDT_L,
  pastDistrib,
  MaxTermGapSize,
  available,
  regr,
  reglog,
  k,
  npt,
  noise
)
}
\description{
Imputation of the terminal created model
}
