% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{addcluster}
\alias{addcluster}
\title{Function that adds the clustering result to an imputed dataset
obtained with seqimpute}
\usage{
addcluster(imputed, clustering)
}
\arguments{
\item{imputed}{a dataframe obtained with the seqimpute function}

\item{clustering}{clustering made on the stacked on multiple imputed datasets}
}
\description{
Function that adds the clustering result to an imputed dataset
obtained with seqimpute
}
