% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICommon.R
\name{FutureVICompute}
\alias{FutureVICompute}
\title{Future VI computation}
\usage{
FutureVICompute(
  CD,
  CO,
  OD,
  ncot,
  frameSize,
  iter,
  nr,
  nc,
  ud,
  np,
  COtsample,
  COt,
  pastDistrib,
  futureDistrib,
  k,
  nf
)
}
\description{
Future VI computation
}
