% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqimpute.R
\name{seqimpute}
\alias{seqimpute}
\title{Imputation of missing data in sequence analysis}
\usage{
seqimpute(
  OD,
  regr = "mlogit",
  np = 1,
  nf = 0,
  nfi = 1,
  npt = 1,
  available = TRUE,
  CO = matrix(NA, nrow = 1, ncol = 1),
  COt = matrix(NA, nrow = 1, ncol = 1),
  pastDistrib = FALSE,
  futureDistrib = FALSE,
  mi = 1,
  mice.return = FALSE,
  include = FALSE,
  noise = 0,
  SetRNGSeed = FALSE,
  ParExec = TRUE,
  num.trees = 10,
  min.node.size = NULL,
  max.depth = NULL,
  timing = F
)
}
\arguments{
\item{OD}{a data frame containing sequences of a multinomial variable with missing data (coded as \code{NA}).}

\item{regr}{a character corresponding to the type of regression model the user want to use for the imputation. The prediction (either multinomial with "\code{mlogit}" or randomForest with "\code{rf}") (default \code{mlogit}).}

\item{np}{\code{numeric} object corresponding to the number of previous observations in the imputation model of the internal gaps (default \code{1}).}

\item{nf}{\code{numeric} object corresponding to the number of future observations in the imputation model of the internal gaps (default \code{0}).}

\item{nfi}{\code{numeric} object corresponding to the number of future observations in the imputation model of the initial gaps (default \code{1}).}

\item{npt}{\code{numeric} object corresponding to the number of previous observations in the imputation model of the terminal gaps (default \code{1}).}

\item{available}{a logical value allowing the user to choose whether to consider the already imputed data in our predictive model (\code{available = TRUE}) or not (\code{available = FALSE}) (default \code{TRUE}).}

\item{CO}{a data frame containing some covariates among which the user can choose in order to specify his model more accurately (default empty matrix 1x1 (\code{matrix(NA,nrow=1,ncol=1)})).}

\item{COt}{a data frame object containing some time-dependent covariates that help specifying the predictive model more accurately (default empty matrix 1x1 (\code{matrix(NA,nrow=1,ncol=1)})).}

\item{pastDistrib}{\code{logical} object allowing to take account of the past distribution in the multinomial logistic regression model or not (default \code{FALSE}).}

\item{futureDistrib}{\code{logical} object allowing to take account of the future distribution in the multinomial logistic regression model or not (default \code{FALSE}).}

\item{mi}{\code{numeric} object corresponding to the number of imputations the program is going to perform (default: \code{1}).}

\item{mice.return}{If \code{TRUE}, an object of class \code{mids}, that can be directly used by the \code{mice} package, is returned.}

\item{include}{\code{logical} object that determines, in the case where a \code{data.frame} is returnes, if the original
dataset should be included or not. This parameter does not apply if \code{mice.return=TRUE}.}

\item{noise}{\code{numeric} object adding a noise on the predicted variable \code{pred} determined by the multinomial model (by introducing a variance \code{noise} for each components of the vector \code{pred}) (the user can choose any value for \code{noise}, but we recommend to choose a rather relatively small value situated in the interval \code{[0.005-0.03]}) (default \code{0}).}

\item{SetRNGSeed}{\code{numeric} This parameter allows for setting RNG seed. Putting a \code{set.seed()} alone before SequImpute function won't work if it is using parallel computation.}

\item{ParExec}{\code{logical} Execute the mi imputation using multicore processing if \code{ParExec=TRUE}. This allows faster run time depending of how many core the processor has. Set \code{ParExec=FALSE} if the processor has 2 or less core.}

\item{num.trees}{\code{integer} Random forest parameter setting the number of trees of each random forest model.}

\item{min.node.size}{\code{interger} Random forest parameter setting the minimum node size for each random forest model.}

\item{max.depth}{\code{interger} Random forest parameter setting the maximal depth tree for each random forest model.}

\item{timing}{\code{logical} indicating whether the place of occurence of the missing data should be added in the imputation model.}
}
\value{
Returns either an S3 object of class \code{mids} if \code{mice.return = TRUE}
or a dataframe, where the imputed dataset are stacked vertically. In the second case,
two columns are added: \code{.imp} integer that refers to the imputation number
(0 corresponding to the original dataset if \code{include=TRUE}) and \code{.id} character corresponding to
the rownames of the dataset to impute.
}
\description{
Multiple imputation of missing data present in a dataset through the prediction based
on either a multinomial or a random forest regression model.
Covariates and time-dependant covariates can be included in the model.
The prediction of the missing values is based on the theory of Prof. Brendan
Halpin. It considers a various amount of surrounding available information to
perform the prediction process.
In fact, we can among others specify \code{np} (the number of past variables
taken into account) and \code{nf} (the number of future information taken
into account).
}
\details{
The imputation process is divided into several steps. According to the location of the gaps of NA among the original dataset, we have defined 5 types of gaps:

- Internal Gaps (simple usual gaps)

- Initial Gaps (gaps situated at the very beginning of a sequence)

- Terminal Gaps (gaps situaed at the very end of a sequence)

- Left-hand side SLG (Specially Located Gaps) (gaps of which the beginning location is included in the interval \code{[0,np]})

- Right-hand side SLG (Specially Located Gaps) (gaps of which the ending location is included in the interval \code{[ncol(OD)-nf,ncol(OD)]})

Order of imputation of the gaps types:
    1. Internal Gaps
    2. Initial Gaps
    3. Terminal Gaps
    4. Left-hand side SLG
    5. Right-hand side SLG
}
\examples{
\dontrun{
# Seqimpute used without parallelisation
RESULT <- seqimpute(OD=OD, np=1, nf=0, nfi=1, npt=1, CO=CO, COt=COt, mi=2)

# Seqimpute used with parallelisation
RESULT <- seqimpute(OD=OD, np=1, nf=0, nfi=1, npt=1, CO=CO, COt=COt, mi=2, ParExec=TRUE)
}
}
\references{
HALPIN, Brendan (2012). Multiple imputation for life-course sequence data. Working Paper WP2012-01, Department of Sociology, University of Limerick. http://hdl.handle.net/10344/3639.

HALPIN, Brendan (2013). Imputing sequence data: Extensions to initial and terminal gaps, Stata's. Working Paper WP2013-01, Department of Sociology, University of Limerick. http://hdl.handle.net/10344/3620
}
\author{
Andre Berchtold <andre.berchtold@unil.ch> Kevin Emery Anthony Guinchard Kamyar Taher
}
