% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqimpute.R
\name{seqimpute}
\alias{seqimpute}
\title{Imputation of missing data in sequence analysis}
\usage{
seqimpute(
  OD,
  regr = "mlogit",
  k,
  np = 1,
  nf = 0,
  nfi = 1,
  npt = 1,
  available = TRUE,
  CO = matrix(NA, nrow = 1, ncol = 1),
  COt = matrix(NA, nrow = 1, ncol = 1),
  pastDistrib = FALSE,
  futureDistrib = FALSE,
  mi = 1,
  mi.return = 1,
  noise = 0
)
}
\arguments{
\item{OD}{\code{matrix} object containing sequences of a multinomial variable with missing data (coded as \code{NA}).}

\item{regr}{\code{character} object corresponding to the type of regression model the user want to use to compute. The prediction (either multinomial with "\code{mlogit}", linear with "\code{lm}" or ordinal with "\code{lrm}") (default \code{mlogit}).}

\item{k}{\code{numeric} object corresponding to the number of categories of the multinomial variable numbered from \code{1} to \code{k}.}

\item{np}{\code{numeric} object corresponding to the number of previous observations in the imputation model of the internal gaps (default \code{1}).}

\item{nf}{\code{numeric} object corresponding to the number of future observations in the imputation model of the internal gaps (default \code{0}).}

\item{nfi}{\code{numeric} object corresponding to the number of future observations in the imputation model of the initial gaps (default \code{1}).}

\item{npt}{\code{numeric} object corresponding to the number of previous observations in the imputation model of the terminal gaps (default \code{1}).}

\item{available}{\code{logical} object allowing the user to choose whether to consider the already imputed data in our predictive model (\code{available = TRUE}) or not (\code{available = FALSE}) (default \code{TRUE}).}

\item{CO}{\code{data.frame} object containing some covariates among which the user can choose in order to specify his model more accurately (default empty matrix 1x1 (\code{matrix(NA,nrow=1,ncol=1)})).}

\item{COt}{\code{data.frame} object containing some time-dependent covariates that help specifying the predictive model more accurately (default empty matrix 1x1 (\code{matrix(NA,nrow=1,ncol=1)})).}

\item{pastDistrib}{\code{logical} object allowing to take account of the past distribution in the multinomial logistic regression model or not (default \code{FALSE}).}

\item{futureDistrib}{\code{logical} object allowing to take account of the future distribution in the multinomial logistic regression model or not (default \code{FALSE}).}

\item{mi}{\code{numeric} object corresponding to the number of imputations the program is going to perform (default: \code{1}).}

\item{mi.return}{\code{numeric} object being either "\code{1}" (omitting \code{OD}) or "\code{2}" (including \code{OD}) and corresponding to the two possible returned formats of the final matrix \code{RESULT} (default: \code{1}).}

\item{noise}{\code{numeric} object adding a noise on the predicted variable \code{pred} determined by the multinomial model (by introducing a variance \code{noise} for each components of the vector \code{pred}) (the user can choose any value for \code{noise}, but we recommend to choose a rather relatively small value situated in the interval \code{[0.005-0.03]}) (default \code{0}).}
}
\value{
RESULT \code{data.frame} object containing the imputed original dataset.
}
\description{
Imputation of missing data present in a dataset through the prediction based
on either a multinomial, a linear or an ordinal regression model.
In order to specify even more the prediction, fixed as well as time-dependant
covariates be included in the model.
The prediction of the missing values is based on the theory of Prof. Brendan
Halpin. It considers a various amount of surrounding available information to
perform the prediction process.
In fact, we can among others specify \code{np} (the number of past variables
taken into account) and \code{nf} (the number of future information taken
into account).
}
\details{
The imputation process is divided into several steps. According to the location of the gaps of NA among the original dataset, we have defined 5 types of gaps:

- Internal Gaps (simple usual gaps)

- Initial Gaps (gaps situated at the very beginning of a sequence)

- Terminal Gaps (gaps situaed at the very end of a sequence)

- Left-hand side SLG (Specially Located Gaps) (gaps of which the beginning location is included in the interval \code{[0,np]})

- Right-hand side SLG (Specially Located Gaps) (gaps of which the ending location is included in the interval \code{[ncol(OD)-nf,ncol(OD)]})

Order of imputation of the gaps types:
    1. Internal Gaps
    2. Initial Gaps
    3. Terminal Gaps
    4. Left-hand side SLG
    5. Right-hand side SLG
}
\examples{
data(OD, CO, COt)

RESULT <- seqimpute(OD=OD, k=2, np=1, nf=0, nfi=1, npt=1, available=TRUE, CO=CO, COt=COt, pastDistrib=FALSE, futureDistrib=FALSE, mi=1, mi.return=1, noise=0)

}
\references{
HALPIN, Brendan, March 2013. Imputing Sequence Data : Extensions to initial and terminal gaps, Stata's mi. Unviversity of Limerick Department of Sociology Working Paper Series. Working Paper WP2013-01, p.3. Available at : http://www.ul.ie/sociology/pubs/wp2013-01.pdf
}
\author{
Andre Berchtold <andre.berchtold@unil.ch>
}
\keyword{data,}
\keyword{logistic}
\keyword{missing}
\keyword{multinomial}
\keyword{patterns}
\keyword{regression,}
