% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{seqmissfplot}
\alias{seqmissfplot}
\title{Plot of the most common patterns of missing data.}
\usage{
seqmissfplot(seqdata, with.complete = TRUE, ...)
}
\arguments{
\item{seqdata}{a state sequence object built with the TraMineR package}

\item{...}{parameters to be passed to the seqfplot function}
}
\description{
Plot of the most common patterns of missing data.
}
