% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{seqmissimplic}
\alias{seqmissimplic}
\title{Function built on the seqimplic function of the TraMineRextras package.
Visualization and identification of the states
that best characterize sequence with missing data vs the sequences without missing data at each position (time point).
See the seqimplic helps to more details on how it works.}
\usage{
seqmissimplic(seqdata, ...)
}
\arguments{
\item{seqdata}{a state sequence object built with the TraMineR package}

\item{...}{parameters to be passed to the seqimplic() function}
}
\description{
Function built on the seqimplic function of the TraMineRextras package.
Visualization and identification of the states
that best characterize sequence with missing data vs the sequences without missing data at each position (time point).
See the seqimplic helps to more details on how it works.
}
