% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{seqmissIplot}
\alias{seqmissIplot}
\title{Plot all the patterns of missing data.}
\usage{
seqmissIplot(seqdata, with.complete = TRUE, ...)
}
\arguments{
\item{seqdata}{a state sequence object built with the TraMineR package}

\item{...}{parameters to be passed to the seqIplot function}
}
\description{
Plot all the patterns of missing data.
}
