% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqQuickLook.R
\name{seqQuickLook}
\alias{seqQuickLook}
\title{Numbering NAs and types of gaps among a dataset}
\usage{
seqQuickLook(OD, k, np = 1, nf = 0)
}
\arguments{
\item{OD}{\code{matrix} object containing sequences of a variable with missing data (coded as \code{NA}).}

\item{k}{\code{numeric} object corresponding to the number of categories of the variable numbered from \code{1} to \code{k}.}

\item{np}{\code{numeric} object corresponding to the number of previous observations in the imputation model of the internal gaps (default \code{1}).}

\item{nf}{\code{numeric} object corresponding to the number of future observations in the imputation model of the internal gaps (default \code{0}).}
}
\value{
It returns a message containing the MDGapsChart \code{data.frame} object that summarizes the number and the
size of each type of gaps (Internal Gaps, Initial Gaps, Terminal Gaps, LEFT-
hand side SLG, RIGHT-hand side SLG) present among the original dataset.
}
\description{
\code{seqQuickLook.R} is a function aimed at providing a quick overview of the
frequency of the NAs and the number and size of the different types of gaps
spread in the original dataset \code{OD}. The user should run this function before
running the main function \code{seqimpute.R} in order to identify a judicious choice
for the values of \code{np} and \code{nf}.
}
\examples{
data(OD)

seqQuickLook(OD=OD, k=2, np=1, nf=0)

}
\author{
Andre Berchtold <andre.berchtold@unil.ch>
}
\keyword{data,}
\keyword{linear}
\keyword{logistic}
\keyword{missing}
\keyword{multinomial}
\keyword{ordinal}
\keyword{patterns}
\keyword{regression,}
