% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqTrans.R
\name{seqTrans}
\alias{seqTrans}
\title{Computing and spotting transitions among a dataset}
\usage{
seqTrans(OD, k, trans)
}
\arguments{
\item{OD}{\code{matrix} object containing sequences of a variable with missing data (coded as \code{NA}).}

\item{k}{\code{numeric} object corresponding to the number of categories of the variable numbered from \code{1} to \code{k}.}

\item{trans}{\code{character} vector gathering the desired transitions. For example: trans <- c("1->3","1->4","2->1","4->1","4->3")}
}
\value{
It returns a message containing the transOverview \code{data.frame} object that gathers the occurences of each type of impossible transitions.

rowMat \code{matrix} object containing the row coordinates of the impossible transitions (stored in \code{seqTransList[1]}).

colMat \code{matrix} object containing the column coordinates of the impossible transitions (stored in \code{seqTransList[2]}).
}
\description{
The purpose of \code{seqTrans.R} is to compute and spot transitions
in a dataset.
}
\examples{
data(OD)

seqTransList <- seqTrans(OD=OD, k=2, trans=c("yes->no"))

}
\author{
Andre Berchtold <andre.berchtold@unil.ch>, Kevin Emery
}
\keyword{data,}
\keyword{impossible}
\keyword{linear}
\keyword{logistic}
\keyword{missing}
\keyword{multinomial}
\keyword{ordinal}
\keyword{regression,}
\keyword{transitions}
