% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{seqwithmiss}
\alias{seqwithmiss}
\title{Extract all the sequences with at least one missing value}
\usage{
seqwithmiss(seqdata)
}
\arguments{
\item{seqdata}{a state sequence object built with the TraMineR package}
}
\value{
Returns a state sequence object containing only
the sequences with at least one missing value
}
\description{
Extract all the sequences with at least one missing value
}
